facial-recognition-attendance/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.6.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
face-recognition==1.3.0
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.9.0.80 